% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_catchment.R
\docType{data}
\name{test_catchment}
\alias{test_catchment}
\title{Example dataset: test_catchment}
\format{
A data frame with 9131 rows and 3 variables:
\describe{
  \item{Date}{Date values from 1990 to 2014}
  \item{precipitation_mmd}{Numeric values of precipitation (rainfall) in mm/day}
  \item{streamflow_mmd}{Numeric values of streamflow in mm/day}
}
}
\source{
Giani et al (2022)
}
\usage{
test_catchment
}
\description{
An example dataset originally used by Giani et al. (2022).
}
\keyword{datasets}
